web/requirements.txt
Flask
gitpython
requests